Bump tablewriter to v1.0.9 (#55)

+++ types_pef_params.go
@@ -52,7 +52,7 @@ func (p PEFConfigParamSelector) String() string {
 		return s
 	}
 
-	return fmt.Sprintf("Unknown (%#02x)", p)
+	return fmt.Sprintf("Unknown (%#02x)", uint8(p))
 }
 
 type PEFConfigParameter interface {
